FCnet added
previously my_en, minor changes in the sense of returning by default less info. es. coefficients will be needed for FCnetLOO but will need to be recalculated again - though with parallelization on the outer loop this will be faster. function is borderline between an internal and something to expose as it fits a model on something for which it has been trained.