cv_Ncomp gained
The main functions gain a cv_Ncomp parameter for internal crossvalidation of the number of components to use in the models. Parameters "nfolds" and "rep_cv" are now discarded to make clear that the functions lead toward the leave-one-out pathway. The default for the minimization procedure is now the mean squared error, yet again.